requirements.txt
cv_bridge==3.2.1
groundingdino_py==0.4.0
huggingface_hub==0.22.2
numpy==1.21.5
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==9.0.1
Pillow==10.3.0
rclpy==3.3.12
rosbag2_py==0.15.9
rosidl_runtime_py==0.9.3
segment_anything==1.0
sensor_msgs==4.2.3
supervision==0.6.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.65.2
tqdm==4.66.2
